+++ sqlparser/lib/src/analysis/types2/graph/type_graph.dart
@@ -231,7 +231,7 @@ class _ResolvedVariables {
   }
 }
 
-extension on ResolvedType {
+extension ResolvedTypeUtils on ResolvedType {
   ResolvedType cast(CastMode mode) {
     switch (mode) {
       case CastMode.numeric: