+++ sqlparser/test/analysis/errors/limit_or_order_by_in_compound_test.dart
@@ -4,13 +4,14 @@ import 'package:test/test.dart';
 void main() {
   test('reports error if LIMIT is used before last part', () {
     final engine = SqlEngine();
-    final analyzed = engine.analyze('SELECT 1 ORDER BY 3 UNION SELECT 2');
+    final analyzed = engine.analyze('SELECT 1 LIMIT 3 UNION SELECT 2');
 
     expect(analyzed.errors, hasLength(1));
     final error = analyzed.errors.single;
 
     expect(error.type, AnalysisErrorType.synctactic);
-    final wrongLimit = (analyzed.root as CompoundSelectStatement).base.orderBy;
+    final problematicSelect = (analyzed.root as CompoundSelectStatement).base;
+    final wrongLimit = (problematicSelect as SelectStatement).limit;
     expect(error.relevantNode, wrongLimit);
   });
 
@@ -26,8 +27,9 @@ void main() {
     final error = analyzed.errors.single;
 
     expect(error.type, AnalysisErrorType.synctactic);
-    final wrongOrderBy =
-        (analyzed.root as CompoundSelectStatement).additional[0].select.orderBy;
+    final problematicSelect =
+        (analyzed.root as CompoundSelectStatement).additional[0].select;
+    final wrongOrderBy = (problematicSelect as SelectStatement).orderBy;
     expect(error.relevantNode, wrongOrderBy);
   });
 }